backend/requirements.txt
fastapi
uvicorn
langchain
langchain-chroma
langchain-openai
langchain-community
langchain-huggingface
chromadb
sentence-transformers
python-dotenv
requests
